Include full HTTP response within 'this' on calls to route handlers and invoke matching route handlers regardless of HTTP response code
Replaced headers, body and fromCache components of the 'this' injected into route handlers with the full HTTP response. This response contains the body and the fromCache aspect is available from the HTTP statusCode that is also part of the response. The body component now comes from the response rather than the body parameter passed into the request.get callback.

Also, removed conditions that prevented route handlers being invoked if the status code was not 200 or 304. Fixed broken tests. Added a new test that uses assert.